cv_project/requirements.txt
streamlit
deepface
opencv-python-headless
numpy
Pillow
streamlit-webrtc
streamlit-autorefresh
google-generativeai
python-dotenv


